+++ nitrocli/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Added the `-v`/`--verbose` option to control libnitrokey log level
+- Added the `-m`/`--model` option to restrict connections to a device
+  model
 
 
 0.2.1